pyproject.toml
[tool.poetry]
name = "imagger"
version = "0.1.0"
description = ""
authors = ["Ilya Torbin <il.torbin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
gunicorn = "^23.0.0"
psycopg2 = "^2.9.10"
pillow = "^11.0.0"
celery = "^5.4.0"
drf-yasg = "^1.21.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
